+++ BookShop/server/graphql/resolvers/shopping.js
@@ -1,42 +1,10 @@
 const Event = require("../../models/event");
-const User = require("../../models/user");
 const Shopping = require("../../models/shopping");
-const { dateToString } = require("../../date");
-const DataLoader = require("dataloader");
 
 const checkAuth = require("../../middleware/is-auth");
 
-const singleEvent = async (id) => {
-  try {
-    const event = await Event.findById(id);
-    return {
-      ...event._doc,
-      id: event.id,
-      // date: dateToString(event._doc.date)
-    };
-  } catch (err) {
-    throw err;
-  }
-};
+const { transformShopping } = require('./merge');
 
-const transformShopping = (shopping) => {
-  return {
-    ...shopping._doc,
-    shoppingId: shopping.id,
-    // user: user.bind(this, shopping._doc.user),
-    event: singleEvent.bind(this, shopping._doc.event),
-    createdAt: dateToString(shopping._doc.createdAt),
-    // updatedAt: dateToString(shopping._doc.updatedAt),
-  };
-};
-
-const transformEvent = (event) => {
-  return {
-    ...event._doc,
-    _id: event.id,
-    date: dateToString(event._doc.date),
-  };
-};
 
 module.exports = {
   Query: {
@@ -45,7 +13,7 @@ module.exports = {
         const { username } = checkAuth(context);
 
         const shoppings = await Shopping.find();
-
+        
         return shoppings
           .filter((shopping) => shopping.username === username)
           .map((shopping) => {
@@ -59,7 +27,10 @@ module.exports = {
   Mutation: {
     async shopEvent(_, args, context) {
       const { username } = checkAuth(context);
+
       const fetchedEvent = await Event.findById({ _id: args.id });
+      if(fetchedEvent === undefined) throw new Error("Event not found");
+
       const shopping = new Shopping({
         event: fetchedEvent,
         username: username,
@@ -68,15 +39,13 @@ module.exports = {
       const result = await shopping.save();
       return transformShopping(result);
     },
-    async cancelShopping(_, args) {
+    async cancelShopping(_, args, context) {
       try {
+        checkAuth(context);
         const shopping = await Shopping.findById(args.shoppingId).populate(
           "event"
         );
-        const event = {
-          ...shopping.event._doc,
-          shoppingId: shopping.event.id,
-        };
+        if(shopping === undefined) throw new Error("Shoppin not found");
         await shopping.delete();
         return "Comment closed successfully";
       } catch (err) {
